+++ src/parser/block/mathblock.ts
@@ -6,11 +6,11 @@ import { html } from 'typed-dom';
 export const segment: MathBlockParser = block(build(() => segment_));
 
 export const segment_: MathBlockParser = block(focus(
-  /^\$\$([^\n]*)(\n(?:[^\n]*\n)*?)\$\$[^\S\n]*(?:\n|$)/,
+  /^\$\$(?!\$)([^\n]*)(\n(?:[^\n]*\n)*?)\$\$[^\S\n]*(?:\n|$)/,
   _ => [[], '']), false);
 
 export const mathblock: MathBlockParser = block(rewrite(segment, match(
-  /^\$\$([^\n]*)(\n(?:[^\n]*\n)*?)\$\$\s*$/,
+  /^\$\$(?!\$)([^\n]*)(\n(?:[^\n]*\n)*?)\$\$\s*$/,
   ([, arg, body], rest) => {
     const el = html('div', { class: `math notranslate` }, `$$${body}$$`);
     if (arg.trim() !== '') {