+++ src/parser/inline/code.ts
@@ -1,9 +1,12 @@
 ﻿import { CodeParser } from '../inline';
-import { match, subline } from '../../combinator';
-import '../source/unescapable';
+import { union, some, match, subline, focus } from '../../combinator';
+import { unescsource } from '../source/unescapable';
 import { html } from 'typed-dom';
 
-export const code: CodeParser = subline(match(
-  /^(`+)(?!`)([^\n]*?[^`\n])\1(?!`)/,
-  ([whole, , body], rest) =>
-    [[html('code', { 'data-src': whole }, body.trim() || body)], rest]));
+export const code: CodeParser = subline(union([
+  match(
+    /^(`+)(?!`)([^\n]*?[^`\n])\1(?!`)/,
+    ([whole, , body], rest) =>
+      [[html('code', { 'data-src': whole }, body.trim() || body)], rest]),
+  focus(/^`+/, some(unescsource)),
+]));